requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
crispy-bulma==0.11.0
cryptography==43.0.1
Django==5.1.1
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-crispy-forms==2.3
django-filter==24.3
django-image-uploader-widget==1.0.0
django-rosetta==0.10.0
djangorestframework==3.15.2
idna==3.10
Markdown==3.7
pillow==10.4.0
polib==1.2.0
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.22
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
